refactor: change dir name of infrastructure to o11y
